Refactor namespaces & directory structure
