+++ src/chesspp/ResourceManager.hpp
