+++ src/chesspp/board/Rook.hpp
@@ -10,13 +10,16 @@
 
 namespace chesspp
 {
-    class Rook : public Piece
+    namespace board
     {
-    private:
-    public:
-        Rook(const Position& bPos, Color c);
+        class Rook : public Piece
+        {
+        private:
+        public:
+            Rook(const Position& bPos, Color c);
 
-        virtual void makeTrajectory(const Board* board);
-    };
+            virtual void makeTrajectory(const Board* board);
+        };
+    }
 }
 #endif
