+++ src/chesspp/graphics/Graphics.hpp
@@ -1,11 +1,11 @@
 #ifndef _GRAPHICS_H
 #define _GRAPHICS_H
 
-#include "SFML.hpp"
-#include "TextureManager.hpp"
-#include "Configuration.hpp"
-#include "board/Piece.hpp"
-#include "board/logger.hpp"
+#include "../../SFML.hpp"
+#include "../TextureManager.hpp"
+#include "../config/Configuration.hpp"
+#include "../board/Piece.hpp"
+#include "../log/Logger.hpp"
 
 #ifdef _DEBUG
 #include <iostream>
@@ -32,16 +32,16 @@ namespace chesspp
             void drawBackground();
 
             //Draws a piece of p.Type and p.Color to p.getBoardPos
-            void drawPiece(Piece *p);
+            void drawPiece(board::Piece *p);
 
             //Separate version of drawPiece to draw a piece at any location on the screen.
-            void drawPieceAt(Piece *p, const sf::Vector2i &pos);
+            void drawPieceAt(board::Piece *p, const sf::Vector2i &pos);
 
             //Draws a valid move block at cell (x,y)
             void drawValidMove(const int x, const int y);
 
             //Draws the entire game.
-            void drawBoard(const Board *b);
+            void drawBoard(const board::Board *b);
         };
     }
 }
